fix: int 오버플로우 방지

+++ 8-period/java-lotto-8-main/src/main/java/lotto/domain/EarningRateCalculator.java
@@ -14,7 +14,7 @@ public class EarningRateCalculator {
     public double calculateEarningRate(int purchasedAmount) {
         double prizeAmount = 0;
         for (Map.Entry<Rank, Integer> entry : rankStatistics.getStatistics().entrySet()) {
-            prizeAmount += entry.getKey().getPrizeAmount() * entry.getValue();
+            prizeAmount += (double) entry.getKey().getPrizeAmount() * entry.getValue();
         }
         return prizeAmount / purchasedAmount * PERCENTAGE;
     }
